Added smartRecordClient and smartRecordServer
