+++ protocol/example/go.mod
@@ -3,7 +3,7 @@ module github.com/libp2p/go-smart-record/protocol/example
 go 1.16
 
 require (
-	github.com/libp2p/go-libp2p v0.13.1-0.20210414131130-b20eb7de2330
+	github.com/libp2p/go-libp2p v0.13.0
 	github.com/libp2p/go-libp2p-core v0.8.5
 	github.com/libp2p/go-smart-record v0.0.0-20210414084730-a9c6e583d4c7
 )
